+++ app/src/main/java/com/example/mygame/player/common/AbstractPlayer.kt
@@ -3,7 +3,7 @@ package com.example.mygame.player.common
 import com.example.mygame.figure.common.FigureColor
 
 abstract class AbstractPlayer : PlayerInterface {
-    abstract var time: Float
+    abstract var time: Long
     abstract var color: FigureColor
     abstract val rock: Int
 }
